// src/interfaces/Propiedad.ts
export interface Propiedad {
  id?: string;
  titulo: string;
  direccion: string;
  calle: string;
  descripcion: string;
  habitaciones: number;
  precio: number;
  wc: number;
  expensas: number;
  barrio: string;
  operacion: string;
  cochera: string;
  tipo: string;
  imagenes: string[];
  amenities: string[];
}